requirements.txt
playwright==1.40.0
redis==5.0.1
google-api-python-client==2.110.0
google-auth==2.25.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
APScheduler==3.10.4
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3